# public/swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "Types.ObjectId": {
        "type": "string"
      },
      "IEvent": {
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "dateLimitBuy": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "nTickets": {
            "type": "number",
            "format": "double"
          },
          "imageUrl": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/Types.ObjectId"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "category",
          "date",
          "dateLimitBuy",
          "description",
          "nTickets",
          "imageUrl",
          "user",
          "price",
          "city",
          "address"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "IPublishEvent": {
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "dateLimitBuy": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "nTickets": {
            "type": "number",
            "format": "double"
          },
          "imageUrl": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "city": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "category",
          "date",
          "dateLimitBuy",
          "description",
          "nTickets",
          "imageUrl",
          "price",
          "city",
          "address"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ITicket": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/Types.ObjectId"
          },
          "event": {
            "$ref": "#/components/schemas/Types.ObjectId"
          },
          "order": {
            "$ref": "#/components/schemas/Types.ObjectId"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["user", "event", "order", "purchaseDate", "price", "date"],
        "type": "object",
        "additionalProperties": false
      },
      "IOrder": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/Types.ObjectId"
          },
          "event": {
            "$ref": "#/components/schemas/Types.ObjectId"
          },
          "nTickets": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "commerceOrder": {
            "type": "string"
          },
          "isPending": {
            "type": "boolean"
          }
        },
        "required": [
          "user",
          "event",
          "nTickets",
          "amount",
          "currency",
          "commerceOrder",
          "isPending"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["message"],
        "type": "object",
        "additionalProperties": false
      },
      "LoginBody": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"],
        "type": "object",
        "additionalProperties": false
      },
      "UserCreateRequest": {
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "secretKey": {
            "type": "string"
          }
        },
        "required": ["username", "email", "password", "role"],
        "type": "object",
        "additionalProperties": false
      },
      "UserResponse": {
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": ["username", "email", "password", "role"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "EntradApp",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "contact": {}
  },
  "openapi": "3.0.0",
  "paths": {
    "/event": {
      "post": {
        "operationId": "CreateEvent",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IEvent"
                }
              }
            }
          }
        },
        "tags": ["event"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IPublishEvent"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "GetEvents",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IEvent"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["event"],
        "security": [],
        "parameters": []
      }
    },
    "/login": {
      "post": {
        "operationId": "LoginUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "tags": ["login"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        }
      }
    },
    "/user": {
      "post": {
        "operationId": "CreateUser",
        "description": "Registra un nuevo usuario en la base de datos.\n Nota: Solamente se pueden registrar correos hotmail, gmail y yahoo.",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateRequest"
                }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRequest"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteUser",
        "description": "Se cambia el estado del usuario cliente a inactivo solamente si no tiene tickets vigentes. Se borra la cuenta del usuario productor solamente si no tiene eventos vigentes.",
        "responses": {
          "202": {
            "description": "- Usuario cliente inactivo\n - Usuario productor borrado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario eliminado"
                    }
                  },
                  "required": ["message"],
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "- Usuario productor tiene eventos vigentes \n- Usuario cliente tiene tickets vigentes \n",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Usuario tiene tickets vigentes"
                    }
                  },
                  "required": ["message"],
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/event/{id}": {
      "get": {
        "operationId": "GetEventById",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IEvent"
                }
              }
            }
          },
          "400": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Event not found"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["event"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "GetUserById",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["user"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/prod/myevents": {
      "get": {
        "operationId": "GetProducerEvents",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IEvent"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["producer"],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/profile": {
      "get": {
        "operationId": "GetProfile",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["profile"],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/order/": {
      "post": {
        "operationId": "PostOrder",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IOrder"
                }
              }
            }
          },
          "400": {
            "description": "No user found\nNo event found\nNo tickets available",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "No tickets available"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["order"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventId": {
                    "type": "string"
                  },
                  "nTickets": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "operationId": "GetOrders",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IOrder"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["order"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": []
      }
    },
    "/order/result/": {
      "post": {
        "operationId": "PostOrderResult",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Transaction successful"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Order not found \nEvent not found \nPending transaction\nRejected\nVoid transaction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order not found"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["order"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/order/paymentConfirm/": {
      "post": {
        "operationId": "PostOrderPaymentConfirm",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Transaction successful"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Order not found \nEvent not found \nPending transaction\nRejected\nVoid transaction",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order not found"
                    }
                  },
                  "required": ["message"],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": ["order"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ticket/": {
      "get": {
        "operationId": "GetTickets",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ITicket"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["ticket"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": []
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Server"
    },
    {
      "url": "https://entradapp-backend.herokuapp.com/api/v1",
      "description": "Heroku Server"
    }
  ]
}